requirements.txt
pillow
pillow-heif
pymediainfo
tqdm
nicegui
pywebview
